+++ inst/resources/server.js
@@ -29,8 +29,36 @@
   chapters.forEach(el => {
     const u = el.dataset.source;
     u && !el.querySelector('.pencil') &&
-      el.insertAdjacentHTML('afterbegin', `<a href="?path=${u}" title="Open ${u}">✎</a>`);
+      el.insertAdjacentHTML('afterbegin', `<span class="buttons"><a href="?path=${u}" title="Open ${u}">✎</a></span>`);
   });
+  const nav = d.querySelector('.nav-path, .title h1');
+  const btn = nav.querySelector('.buttons') || d.createElement('span');
+  btn.className = 'buttons';
+  ['back', 'forward', 'refresh', 'print'].forEach((action, i) => {
+    if (btn.querySelector(`.${action}`)) return;
+    const a = d.createElement('a');
+    a.href = '#'; a.title = action[0].toUpperCase() + action.slice(1);
+    const k = ({
+      back: 'Alt + LeftArrow', forward: 'Alt + RightArrow',
+      refresh: 'Ctrl + R / Command + R'
+    })[action];
+    if (k) a.title += ` (${k})`;
+    a.className = action + ' btn-lite';
+    a.innerText = ['←', '→', '⟳', '⎙'][i];
+    a.onclick = e => btnAction(e, action);
+    btn.append(a);
+  });
+  if (nav) {
+    nav.querySelectorAll('a.btn-lite').forEach(a => btn.append(a));
+    nav.append(btn);
+  }
+  function btnAction(e, action) {
+    if (!action) return;
+    e.preventDefault();
+    action === 'print' ? window.print() : (
+      action === 'refresh' ? location.reload() : history[action]()
+    );
+  }
   // add classes and events to edit buttons
   d.querySelectorAll('a[href]').forEach(a => {
     if (a.innerText !== '✎' || a.classList.contains('pencil')) return;
@@ -41,9 +69,9 @@
     };
   });
   // add classes and events to save buttons
-  d.querySelectorAll('a[href]').forEach(a => {
-    if (a.innerText !== '↯' || a.classList.contains('save')) return;
-    a.classList.add('save'); a.title = 'Render and Save';
+  d.querySelectorAll('a.save[href]').forEach(a => {
+    if (a.innerText !== '↯') return;
+    a.title = 'Render to disk';
     a.onclick = e => {
       e.preventDefault();
       const cls = d.body.classList;
@@ -101,7 +129,7 @@
     });
     // update cross-references
     w.querySelectorAll('a[class^="cross-ref-"][href^="#"]').forEach(el => {
-      const n = d.querySelector(el.getAttribute('href'))
+      const n = d.getElementById(el.getAttribute('href').replace(/^#/, ''))
         ?.querySelector('.section-number,[class^="ref-number-"]')
         ?.innerText;
       if (n) el.innerText = n;
@@ -146,7 +174,15 @@
       };
     });
   }
-  window.addEventListener('load', new_interval);
+  window.addEventListener('load', () => {
+    d.onkeydown = e => {
+      const k = e.key; let a;
+      k === 'r' && (e.metaKey || e.ctrlKey) && (a = 'refresh');
+      e.altKey && (a = ({'ArrowLeft': 'back', 'ArrowRight': 'forward'})[k]);
+      btnAction(e, a);
+    };
+    new_interval();
+  });
   // when updating a book chapter, this script will be reloaded, and we need to
   // clear the old interval and create a new loop
   if (d.body.dataset.timerId) {